+++ pkg/providers/papi/create_property_test.go
@@ -530,6 +530,9 @@ func TestCreateProperty(t *testing.T) {
 		VersionStaging:          "1",
 	}
 
+	var noFilters []func([]string) ([]string, error)
+	otherRuleFormatFilter := []func([]string) ([]string, error){useThisOnlyRuleFormat("v2023-01-05")}
+
 	tests := map[string]struct {
 		init                func(*papi.Mock, *hapi.Mock, *templates.MockProcessor, string)
 		dir                 string
@@ -556,7 +559,7 @@ func TestCreateProperty(t *testing.T) {
 				mockGetEdgeHostnames(c)
 				mockGetActivations(c, &getActivationsResponse, nil)
 				mockGetActivations(c, &papi.GetActivationsResponse{}, nil)
-				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().build(), nil)
+				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().build(), noFilters, nil)
 			},
 			dir:     "basic",
 			jsonDir: "basic/property-snippets",
@@ -586,7 +589,7 @@ func TestCreateProperty(t *testing.T) {
 				mockGetActivations(c, &getActivationsResponse, nil)
 				mockGetActivations(c, &papi.GetActivationsResponse{}, nil)
 				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withEdgeHostname(map[string]EdgeHostname{}).
-					withIsActive(false).build(), nil)
+					withIsActive(false).build(), noFilters, nil)
 			},
 			dir:     "basic_property_with_empty_hostname_id",
 			jsonDir: "basic_property_with_empty_hostname_id/property-snippets",
@@ -616,7 +619,7 @@ func TestCreateProperty(t *testing.T) {
 				mockAddTemplateTargetRules(p)
 				mockTemplateExist(p, "rules_v2023-01-05.tmpl", true)
 				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withRuleFormat("v2023-01-05").
-					withRules(flattenRules("test.edgesuite.net", ruleResponse.Rules)).build(), nil)
+					withRules(flattenRules("test.edgesuite.net", ruleResponse.Rules)).build(), otherRuleFormatFilter, nil)
 			},
 			dir:    "basic-rules-datasource",
 			schema: true,
@@ -640,7 +643,7 @@ func TestCreateProperty(t *testing.T) {
 				mockAddTemplateTargetRules(p)
 				mockTemplateExist(p, "rules_latest.tmpl", false)
 				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withIsActive(false).
-					withRules(flattenRules("test.edgesuite.net", ruleResponse.Rules)).build(), nil)
+					withRules(flattenRules("test.edgesuite.net", ruleResponse.Rules)).build(), noFilters, nil)
 			},
 			withError: ErrUnsupportedRuleFormat,
 			dir:       "basic",
@@ -665,7 +668,7 @@ func TestCreateProperty(t *testing.T) {
 				mockAddTemplateTargetRules(p)
 				mockTemplateExist(p, "rules_v2023-01-05.tmpl", true)
 				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withRuleFormat("v2023-01-05").
-					withRules(flattenRules("test.edgesuite.net", ruleResponse.Rules)).build(), ErrSavingFiles)
+					withRules(flattenRules("test.edgesuite.net", ruleResponse.Rules)).build(), otherRuleFormatFilter, ErrSavingFiles)
 			},
 			withError: ErrSavingFiles,
 			dir:       "basic-rules-datasource-unknown",
@@ -695,7 +698,7 @@ func TestCreateProperty(t *testing.T) {
 				mockGetEdgeHostnames(c)
 				mockGetActivations(c, &getActivationsResponse, nil)
 				mockGetActivations(c, &papi.GetActivationsResponse{}, nil)
-				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withIncludes([]TFIncludeData{tfIncludeData}).build(), nil)
+				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withIncludes([]TFIncludeData{tfIncludeData}).build(), noFilters, nil)
 			},
 			dir:     "basic_property_with_include",
 			jsonDir: "basic_property_with_include/property-snippets",
@@ -745,7 +748,7 @@ func TestCreateProperty(t *testing.T) {
 				mockGetEdgeHostnames(c)
 				mockGetActivations(c, &getActivationsResponse, nil)
 				mockGetActivations(c, &papi.GetActivationsResponse{}, nil)
-				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withIncludes([]TFIncludeData{tfIncludeData, tfIncludeData1}).build(), nil)
+				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withIncludes([]TFIncludeData{tfIncludeData, tfIncludeData1}).build(), noFilters, nil)
 			},
 			dir:     "basic_property_with_multiple_includes",
 			jsonDir: "basic_property_with_multiple_includes/property-snippets",
@@ -807,7 +810,7 @@ func TestCreateProperty(t *testing.T) {
 					withIncludes([]TFIncludeData{tfIncludeData, tfIncludeData1}).
 					withIncludeRules(0, flattenRules("test_include", includeRuleResponse.Rules)).
 					withIncludeRules(1, flattenRules("test_include_1", secondIncludeRuleResponse.Rules)).
-					build(), nil)
+					build(), otherRuleFormatFilter, nil)
 				mockAddTemplateTargetRules(p)
 				mockTemplateExist(p, "rules_v2023-01-05.tmpl", true)
 			},
@@ -831,7 +834,7 @@ func TestCreateProperty(t *testing.T) {
 				mockGetEdgeHostnames(c)
 				mockGetActivations(c, &getActivationsResponse, nil)
 				mockGetActivations(c, &papi.GetActivationsResponse{}, nil)
-				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withCertProvisioningType("DEFAULT").build(), nil)
+				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withCertProvisioningType("DEFAULT").build(), noFilters, nil)
 			},
 			dir:     "basic_with_cert_provisioning_type",
 			jsonDir: "basic_with_cert_provisioning_type/property-snippets",
@@ -858,7 +861,7 @@ func TestCreateProperty(t *testing.T) {
 				mockGetEdgeHostnames(c)
 				mockGetActivations(c, &getActivationsResponse, nil)
 				mockGetActivations(c, &papi.GetActivationsResponse{}, nil)
-				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().build(), nil)
+				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().build(), noFilters, nil)
 			},
 			dir:     "basic",
 			jsonDir: "basic/property-snippets",
@@ -884,7 +887,7 @@ func TestCreateProperty(t *testing.T) {
 				mockGetEdgeHostnames(c)
 				mockGetActivations(c, &getActivations1Response, nil)
 				mockGetActivations(c, &papi.GetActivationsResponse{}, nil)
-				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withVersion("1").withStagingVersion(1).build(), nil)
+				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withVersion("1").withStagingVersion(1).build(), noFilters, nil)
 			},
 			dir:     "basic-v1",
 			jsonDir: "basic-v1/property-snippets",
@@ -913,7 +916,7 @@ func TestCreateProperty(t *testing.T) {
 				mockGetActivations(c, &getActivationsResponseWithNote, nil)
 				mockGetActivations(c, &papi.GetActivationsResponse{}, nil)
 				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withActivationNote("example staging note").
-					withEmails([]string{"jsmith@akamai.com", "rjohnson@akamai.com"}).build(), nil)
+					withEmails([]string{"jsmith@akamai.com", "rjohnson@akamai.com"}).build(), noFilters, nil)
 			},
 			dir: "basic",
 		},
@@ -932,7 +935,7 @@ func TestCreateProperty(t *testing.T) {
 				mockGetEdgeHostnames(c)
 				mockGetActivations(c, &papi.GetActivationsResponse{}, nil)
 				mockGetActivations(c, &getProductionActivationsResponse, nil)
-				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withOnlyProductionActivation([]string{"jsmith@akamai.com", "rjohnson@akamai.com"}, "example production note", 2).build(), nil)
+				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withOnlyProductionActivation([]string{"jsmith@akamai.com", "rjohnson@akamai.com"}, "example production note", 2).build(), noFilters, nil)
 			},
 			dir: "basic",
 		},
@@ -952,7 +955,7 @@ func TestCreateProperty(t *testing.T) {
 				mockGetActivations(c, &getActivationsResponseWithNote, nil)
 				mockGetActivations(c, &getProductionActivationsResponse, nil)
 				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withOnlyProductionActivation([]string{"jsmith@akamai.com", "rjohnson@akamai.com"}, "example production note", 2).withActivationNote("example staging note").
-					withEmails([]string{"jsmith@akamai.com", "rjohnson@akamai.com"}).withStagingVersion(2).build(), nil)
+					withEmails([]string{"jsmith@akamai.com", "rjohnson@akamai.com"}).withStagingVersion(2).build(), noFilters, nil)
 			},
 			dir: "basic",
 		},
@@ -973,7 +976,7 @@ func TestCreateProperty(t *testing.T) {
 				mockGetActivations(c, &getActivationsResponseWithEmptyEmails, nil)
 				mockGetActivations(c, &papi.GetActivationsResponse{}, nil)
 				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().withActivationNote("example note").
-					withEmails([]string{""}).build(), nil)
+					withEmails([]string{""}).build(), noFilters, nil)
 			},
 			dir: "basic",
 		},
@@ -1091,7 +1094,7 @@ func TestCreateProperty(t *testing.T) {
 				mockGetActivations(c, &getActivationsResponse, nil)
 				mockAddTemplateTargetRules(p)
 				mockTemplateExist(p, "rules_v2023-01-05.tmpl", true)
-				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().build(), fmt.Errorf("oops"))
+				mockProcessTemplates(p, (&tfDataBuilder{}).withDefaults().build(), noFilters, fmt.Errorf("oops"))
 			},
 			dir:       "basic",
 			withError: ErrSavingFiles,
@@ -1244,8 +1247,12 @@ func mockGetIncludeRuleTree(c *papi.Mock, getIncludeRuleTreeReq papi.GetIncludeR
 	c.On("GetIncludeRuleTree", mock.Anything, getIncludeRuleTreeReq).Return(includeRuleResponse, nil).Once()
 }
 
-func mockProcessTemplates(p *templates.MockProcessor, tfData TFData, err error) {
-	p.On("ProcessTemplates", tfData).Return(err).Once()
+func mockProcessTemplates(p *templates.MockProcessor, tfData TFData, filterFuncs []func([]string) ([]string, error), err error) {
+	if len(filterFuncs) != 0 {
+		p.On("ProcessTemplates", tfData, mock.AnythingOfType("func([]string) ([]string, error)")).Return(err).Once()
+	} else {
+		p.On("ProcessTemplates", tfData).Return(err).Once()
+	}
 }
 
 func getRuleTreeResponse(dir string, t *testing.T) papi.GetRuleTreeResponse {
@@ -1277,6 +1284,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		schema       bool
 		ruleResponse papi.GetRuleTreeResponse
 		withError    string
+		filterFuncs  []func([]string) ([]string, error)
 	}{
 		"property": {
 			givenData: TFData{
@@ -1366,6 +1374,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 			dir:          "basic-rules-datasource",
 			schema:       true,
 			filesToCheck: []string{"property.tf", "rules.tf", "variables.tf", "import.sh"},
+			filterFuncs:  []func([]string) ([]string, error){useThisOnlyRuleFormat("v2023-01-05")},
 		},
 		"property with rules as datasource with unknown behaviors and criteria": {
 			givenData: TFData{
@@ -1625,6 +1634,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 			filesToCheck: []string{"property.tf", "includes.tf", "variables.tf", "import.sh", "includes_rules.tf"},
 			withIncludes: true,
 			schema:       true,
+			filterFuncs:  []func([]string) ([]string, error){useThisOnlyRuleFormat("v2023-01-05")},
 		},
 		"property with use cases": {
 			givenData: TFData{
@@ -1856,6 +1866,52 @@ func TestProcessPolicyTemplates(t *testing.T) {
 			dir:          "basic_without_activation",
 			filesToCheck: []string{"property.tf", "variables.tf", "import.sh"},
 		},
+		// property with rules as datasource - schema version 1 and 2 is a pair of tests that confirms that schema 1 does not use any schema's 2 template definitions and vice versa
+		// the behaviour was chosen in a way, so it's easily identifiable which template inner definition was picked (e.g. there was change in field type)
+		"property with rules as datasource - schema version 1": {
+			givenData: TFData{
+				Property: TFPropertyData{
+					GroupName:            "test_group",
+					GroupID:              "grp_12345",
+					ContractID:           "test_contract",
+					PropertyResourceName: "test-edgesuite-net",
+					PropertyName:         "test.edgesuite.net",
+					PropertyID:           "prp_12345",
+					ProductID:            "prd_HTTP_Content_Del",
+					ProductName:          "HTTP_Content_Del",
+					RuleFormat:           "v2023-01-05",
+					IsSecure:             "false",
+					ReadVersion:          "LATEST",
+				},
+				Section: "test_section",
+			},
+			dir:          "basic-rules-datasource-schema1",
+			schema:       true,
+			filesToCheck: []string{"property.tf", "rules.tf", "variables.tf", "import.sh"},
+			filterFuncs:  []func([]string) ([]string, error){useThisOnlyRuleFormat("v2023-01-05")},
+		},
+		"property with rules as datasource - schema version 2": {
+			givenData: TFData{
+				Property: TFPropertyData{
+					GroupName:            "test_group",
+					GroupID:              "grp_12345",
+					ContractID:           "test_contract",
+					PropertyResourceName: "test-edgesuite-net",
+					PropertyName:         "test.edgesuite.net",
+					PropertyID:           "prp_12345",
+					ProductID:            "prd_HTTP_Content_Del",
+					ProductName:          "HTTP_Content_Del",
+					RuleFormat:           "v2023-05-30",
+					IsSecure:             "false",
+					ReadVersion:          "LATEST",
+				},
+				Section: "test_section",
+			},
+			dir:          "basic-rules-datasource-schema2",
+			schema:       true,
+			filesToCheck: []string{"property.tf", "rules.tf", "variables.tf", "import.sh"},
+			filterFuncs:  []func([]string) ([]string, error){useThisOnlyRuleFormat("v2023-05-30")},
+		},
 	}
 
 	for name, test := range tests {
@@ -1876,7 +1932,13 @@ func TestProcessPolicyTemplates(t *testing.T) {
 				templateToFile["includes.tmpl"] = fmt.Sprintf("./testdata/res/%s/includes.tf", test.dir)
 			}
 			if test.schema {
-				templateToFile["rules_v2023-01-05.tmpl"] = fmt.Sprintf("./testdata/res/%s/rules.tf", test.dir)
+				var rulesVersion string
+				if len(test.givenData.Includes) > 0 {
+					rulesVersion = test.givenData.Includes[0].RuleFormat
+				} else {
+					rulesVersion = test.givenData.Property.RuleFormat
+				}
+				templateToFile[fmt.Sprintf("rules_%s.tmpl", rulesVersion)] = fmt.Sprintf("./testdata/res/%s/rules.tf", test.dir)
 			}
 			if test.withIncludes && test.schema {
 				templateToFile["includes_rules.tmpl"] = fmt.Sprintf("./testdata/res/%s/includes_rules.tf", test.dir)
@@ -1887,7 +1949,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 				TemplateTargets: templateToFile,
 				AdditionalFuncs: additionalFuncs,
 			}
-			err := processor.ProcessTemplates(test.givenData)
+			err := processor.ProcessTemplates(test.givenData, test.filterFuncs...)
 			reportedErrors = []string{}
 			if test.withError != "" {
 				assert.Error(t, err)
@@ -2267,4 +2329,40 @@ func (t *tfDataBuilder) withIncludes(includes []TFIncludeData) *tfDataBuilder {
 
 func (t *tfDataBuilder) build() TFData {
 	return t.tfData
+}
+
+func TestUseThisOnlyRuleFormat(t *testing.T) {
+	tests := map[string]struct {
+		acceptedFormat string
+		input          []string
+		expected       []string
+		err            error
+	}{
+		"happy case": {
+			acceptedFormat: "v2023-05-30",
+			input:          []string{"import.sh", "templates/rules_v2023-01-05.tmpl", "templates/rules_v2023-05-30.tmpl"},
+			expected:       []string{"import.sh", "templates/rules_v2023-05-30.tmpl"},
+		},
+		"missing template": {
+			acceptedFormat: "v2023-05-32",
+			input:          []string{"import.sh", "templates/rules_v2023-01-05.tmpl", "templates/rules_v2023-05-30.tmpl"},
+			err:            fmt.Errorf("did not find v2023-05-32 format among [import.sh templates/rules_v2023-01-05.tmpl templates/rules_v2023-05-30.tmpl]"),
+		},
+		"no formats at all": {
+			acceptedFormat: "v2023-05-32",
+			input:          []string{"import.sh", "READ.md"},
+			err:            fmt.Errorf("did not find v2023-05-32 format among [import.sh READ.md]"),
+		},
+	}
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			output, err := useThisOnlyRuleFormat(test.acceptedFormat)(test.input)
+			if test.err != nil {
+				assert.Equal(t, test.err, err)
+				return
+			}
+			require.NoError(t, err)
+			assert.Equal(t, test.expected, output)
+		})
+	}
 }